+++ .idea/libraries/Dart_SDK.xml
@@ -1,0 +1,29 @@
+<component name="libraryTable">
+  <library name="Dart SDK">
+    <CLASSES>
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/async" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/cli" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/collection" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/convert" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/core" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/developer" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/ffi" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/html" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/indexed_db" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/io" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/isolate" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/js" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/js_util" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/math" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/mirrors" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/svg" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/typed_data" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/wasm" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/web_audio" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/web_gl" />
+      <root url="file://D:/src/flutter/bin/cache/dart-sdk/lib/web_sql" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
